molmoact2: organize demos/scripts, default fast demos, README GIFs
Move package files into demos/, scripts/, assets/. Default both interactive
demos to the fast path (depth off, 4 denoising steps); full depth (THINK=1) and
long horizon (RT_HORIZON_MULT) stay optional. Add interactive/DROID/LIBERO GIFs
and rewrite the README around the demos. Trim comments and drop the unused FAST
knob so a clean build exposes every demo with no extra steps.
