+++ packages/vla/molmoact2/scripts/interactive_server_rt.py
@@ -1,24 +1,18 @@
 # Copyright(C) 2026 Advanced Micro Devices, Inc. All rights reserved.
 # SPDX-License-Identifier: MIT
-"""Interactive MolmoAct2 x LIBERO demo (local inference, browser UI).
+"""Real-time MolmoAct2 + LIBERO demo: the sim keeps running while the model thinks.
 
-Command-driven: the sim sits IDLE on a scene until you send an instruction; then it
-resets the env + policy and runs that one task to completion (or until you hit Stop),
-streaming the FHD-equivalent camera to the browser as MJPEG and saving a debug video
-(executed command banner on top). Randomize swaps to a new random scene.
+Real-time sibling of interactive_server.py. Planning is decoupled from the simulator so
+the planner latency is visible:
+  * a sim thread steps MuJoCo at wall-clock RT_HZ, consuming one action per tick from a
+    shared buffer; when the buffer is empty it applies a HOLD action so the arm stays put.
+  * a planner thread runs the policy forward to refill the buffer with the next chunk.
 
-Why reset on every command: the MolmoAct2-Think policy caches its depth/spatial plan
-once per episode (and runs a receding-horizon action queue), so a new instruction only
-takes effect after `policy.reset()` + a fresh `env.reset()`. Resetting per command is
-both the desired UX and the fix for "it keeps doing the old task".
+HOLD is zeros on the 6 delta dims (OSC_POSE control_delta=True holds the current pose) and
+keeps the last gripper command. Reuses the lerobot select_action path; only the loop is local.
 
-Reuses the validated lerobot eval machinery verbatim (make_env / make_policy /
-processors / rollout); only the loop is ours (live instruction + streaming + video).
-Stdlib http.server only (no extra deps).
-
-Env: SUITE, TASK_ID, SEED, THINK (1), CKPT, PORT (8080), VIEW_RES (1080),
-VIDEO_RES (600), OUT_DIR (/outputs). Open http://localhost:PORT (remote box:
-ssh -L PORT:localhost:PORT <host>).
+Env: SUITE, TASK_ID, SEED, THINK (1), CKPT, PORT (8081), VIEW_RES (720), VIDEO_RES (600),
+RT_HZ (20), RT_LOOKAHEAD (0), OUT_DIR (/outputs).
 """
 import io
 import json
@@ -26,6 +20,7 @@ import os
 import random
 import threading
 import time
+from collections import deque
 from datetime import datetime
 from http.server import BaseHTTPRequestHandler, ThreadingHTTPServer
 from urllib.parse import parse_qs, urlparse
@@ -38,31 +33,34 @@ TASK_ID = int(os.environ.get("TASK_ID", "3"))
 SEED = int(os.environ.get("SEED", "1000"))
 THINK = os.environ.get("THINK", "1") == "1"
 CKPT = os.environ.get("CKPT", "allenai/MolmoAct2-Think-LIBERO")
-PORT = int(os.environ.get("PORT", "8080"))
-VIEW_RES = int(os.environ.get("VIEW_RES", "1080"))   # live viewport (FHD-equivalent)
+PORT = int(os.environ.get("PORT", "8081"))
+VIEW_RES = int(os.environ.get("VIEW_RES", "720"))    # live viewport (RT favors smoothness)
 VIDEO_RES = int(os.environ.get("VIDEO_RES", "600"))  # saved debug video (kept small)
+RT_HZ = float(os.environ.get("RT_HZ", "20"))         # wall-clock control rate
+RT_LOOKAHEAD = int(os.environ.get("RT_LOOKAHEAD", "0"))  # replan when buffer <= this (pipelining knob)
+RT_MAX_STEPS = int(os.environ.get("RT_MAX_STEPS", "1200"))
+# Optional: multiply the simulator's real horizon (robosuite) to play around with
+# long-horizon tasks without the scene resetting mid-execution. Default 1 (unchanged).
+RT_HORIZON_MULT = float(os.environ.get("RT_HORIZON_MULT", "1"))  # RT runs in wall time; holds burn budget, so allow completion
 OUT_DIR = os.environ.get("OUT_DIR", "/outputs")
 SUITES = ["libero_object", "libero_goal", "libero_spatial", "libero_10"]
+DT = 1.0 / RT_HZ
 
 # ---- shared state ------------------------------------------------------------
 STATE = {
     "mode": "loading",        # loading | idle | running
-    "instruction": "",        # the instruction currently being executed
-    "scene_task": "",         # the scene's native LIBERO instruction
+    "instruction": "",
+    "scene_task": "",
     "suite": SUITE, "task_id": TASK_ID,
-    "objects": [],            # object names visible in the scene
+    "objects": [],
     "step": 0, "infer_ms": 0.0, "success": False,
+    "holding": False, "buffer": 0, "hold_pct": 0.0,
     "status": "starting", "frame": None, "video_url": "",
 }
 LOCK = threading.Lock()
-# command mailbox (latest wins): action in {"run","randomize",None}
 PENDING = {"action": None, "instruction": ""}
 EVENT = threading.Event()
 STOP = {"flag": False}
-
-
-class StopRollout(Exception):
-    pass
 
 
 def _font(size):
@@ -73,7 +71,7 @@ def _font(size):
     return ImageFont.load_default()
 
 
-def _encode_jpeg(rgb, quality=88):
+def _encode_jpeg(rgb, quality=85):
     buf = io.BytesIO()
     Image.fromarray(np.ascontiguousarray(rgb)).save(buf, format="JPEG", quality=quality)
     return buf.getvalue()
@@ -84,24 +82,24 @@ def _set_frame(rgb):
         STATE["frame"] = _encode_jpeg(rgb)
 
 
-def _banner_frame(rgb, text, size):
-    """Downscale to `size` and add a top banner with the executed command."""
+def _banner_frame(rgb, text, size, thinking):
+    """Downscale to `size`; top banner shows the command + a THINKING tag when holding."""
     img = Image.fromarray(np.ascontiguousarray(rgb)).resize((size, size), Image.BILINEAR)
     bh = max(40, size // 12)
     bh += bh % 2
     canvas = Image.new("RGB", (size, size + bh), (15, 15, 18))
     canvas.paste(img, (0, bh))
     d = ImageDraw.Draw(canvas)
-    f = _font(max(14, size // 28))
-    msg = text if len(text) <= 70 else text[:67] + "..."
+    f = _font(max(14, size // 30))
+    cap = 44 if thinking else 60  # leave room on the right for the THINKING tag
+    msg = text if len(text) <= cap else text[:cap - 3] + "..."
     d.text((10, bh // 2), msg, fill=(240, 240, 240), font=f, anchor="lm")
+    if thinking:
+        d.text((size - 10, bh // 2), "THINKING", fill=(255, 180, 80), font=f, anchor="rm")
     return np.asarray(canvas)
 
 
 class Scene:
-    """Holds a single (suite, task_id) vec env + its metadata. Built via make_env so
-    obs_type / processors match the validated eval path exactly."""
-
     def __init__(self, suite, task_id, env, scene_task, objects):
         self.suite, self.task_id = suite, task_id
         self.env, self.scene_task, self.objects = env, scene_task, objects
@@ -122,7 +120,7 @@ class Scene:
         return self.hi_res(size)
 
 
-def rollout_thread():
+def engine_thread():
     import torch
     import draccus
     from lerobot.configs.eval import EvalPipelineConfig
@@ -130,11 +128,8 @@ def rollout_thread():
     from lerobot.policies.factory import make_policy, make_pre_post_processors
     import lerobot.scripts.lerobot_eval as ev
 
-    # --- live instruction injection (replaces the scene's fixed task language) ---
-    def _inject(env, observation):
-        observation["task"] = [STATE["instruction"] for _ in range(env.num_envs)]
-        return observation
-    ev.add_envs_task = _inject
+    ACTION = ev.ACTION
+    preprocess_observation = ev.preprocess_observation
 
     depth = "True" if THINK else "False"
     base_args = [
@@ -144,7 +139,7 @@ def rollout_thread():
         "--policy.enable_cuda_graph=False", "--policy.norm_tag=libero",
         "--policy.device=cuda", "--env.type=libero",
         "--eval.batch_size=1", "--eval.n_episodes=1",
-        f"--seed={SEED}", "--output_dir=/tmp/interactive_eval",
+        f"--seed={SEED}", "--output_dir=/tmp/interactive_rt_eval",
     ]
     if os.environ.get("NUM_STEPS"):
         base_args.append(f"--policy.num_steps={os.environ['NUM_STEPS']}")
@@ -167,6 +162,26 @@ def rollout_thread():
         cfg = draccus.parse(EvalPipelineConfig, args=base_args + [f"--env.task={suite}", f"--env.task_ids=[{task_id}]"])
         envs = make_env(cfg.env, n_envs=1, use_async_envs=False, trust_remote_code=cfg.trust_remote_code)
         env = envs[suite][task_id]
+        for e in env.envs:
+            # robosuite's internal horizon is the real terminator (the gym wrapper never
+            # truncates on _max_episode_steps); raise it by RT_HORIZON_MULT so long tasks
+            # do not reset mid-execution. _max_episode_steps caps our own control loop.
+            base_h = RT_MAX_STEPS
+            rs = getattr(getattr(e, "_env", None), "env", None)
+            if rs is not None and getattr(rs, "horizon", None):
+                base_h = int(rs.horizon)
+            new_h = int(base_h * RT_HORIZON_MULT)
+            if rs is not None:
+                try:
+                    rs.horizon = new_h
+                except Exception:
+                    pass
+            e._max_episode_steps = max(RT_MAX_STEPS, new_h)
+        try:
+            print(f"[scene] horizon_mult={RT_HORIZON_MULT} horizon={new_h} "
+                  f"max_steps={max(RT_MAX_STEPS, new_h)}", flush=True)
+        except Exception:
+            pass
         env.reset(seed=SEED)
         le = env.envs[0]
         scene_task = getattr(le, "task_description", "") or ""
@@ -179,59 +194,146 @@ def rollout_thread():
 
     def show_idle(sc):
         with LOCK:
-            STATE.update(mode="idle", status="idle - send an instruction", step=0,
+            STATE.update(mode="idle", status="idle, send an instruction", step=0,
                          success=False, suite=sc.suite, task_id=sc.task_id,
                          scene_task=sc.scene_task, objects=sc.objects,
-                         instruction="", video_url="")
+                         instruction="", video_url="", holding=False, buffer=0, hold_pct=0.0)
         STATE.pop("_video_path", None)
         _set_frame(sc.idle_frame(VIEW_RES))
 
-    os.makedirs(os.path.join(OUT_DIR, "interactive"), exist_ok=True)
+    os.makedirs(os.path.join(OUT_DIR, "interactive_rt"), exist_ok=True)
     scene = build_scene(SUITE, TASK_ID)
-    show_idle(scene)
+
+    # ---- one planning step: model forward -> full env-space chunk -------------
+    def plan_chunk(proc_obs, generator):
+        with torch.inference_mode():
+            a0 = policy.select_action(proc_obs, generator=generator)  # forward + pop 1
+        raw = [a0]
+        q = policy._action_queues[0]
+        while q:                                                      # drain the rest
+            a = q.popleft()
+            if a.ndim == 1:
+                a = a.unsqueeze(0)
+            raw.append(a.to(device=a0.device, dtype=torch.float32))
+        out = []
+        for a in raw:
+            pa = postprocessor(a)
+            pa = env_post({ACTION: pa})[ACTION]
+            out.append(pa.to("cpu").numpy())
+        infer_s = float(getattr(policy, "_last_model_inference_s", 0.0))
+        print(f"[plan] infer={infer_s:.2f}s chunk={len(out)}", flush=True)
+        return out, infer_s
 
     def run_command(sc, instruction):
         with LOCK:
             STATE.update(mode="running", instruction=instruction, step=0, success=False,
-                         status=f"running: {instruction}", video_url="")
+                         status=f"running: {instruction}", video_url="", holding=False,
+                         buffer=0, hold_pct=0.0)
         STOP["flag"] = False
-        frames = []  # (banner) frames for the saved video
+        policy.reset()
+        obs, _ = sc.env.reset(seed=SEED)
+        generator = ev._make_rollout_action_generator(policy, [SEED])
+        max_steps = sc.env.call("_max_episode_steps")[0]
 
-        def render_cb(vec_env):
-            if STOP["flag"]:
-                raise StopRollout
+        buffer = deque()
+        buf_lock = threading.Lock()
+        shared = {"obs": obs, "done": False, "last_gripper": 0.0,
+                  "hold_steps": 0, "total_steps": 0}
+
+        def planner():
+            while not STOP["flag"] and not shared["done"]:
+                with buf_lock:
+                    have = len(buffer)
+                if have > RT_LOOKAHEAD:
+                    time.sleep(0.005)
+                    continue
+                with LOCK:
+                    cur = shared["obs"]
+                try:
+                    proc = preprocess_observation(cur)
+                    proc["task"] = [instruction for _ in range(sc.env.num_envs)]
+                    proc = env_pre(proc)
+                    proc = preprocessor(proc)
+                    chunk, infer_s = plan_chunk(proc, generator)
+                except Exception as e:  # noqa: BLE001
+                    print("plan error:", e, flush=True)
+                    time.sleep(0.02)
+                    continue
+                with buf_lock:
+                    buffer.extend(chunk)
+                with LOCK:
+                    if infer_s > 0:
+                        STATE["infer_ms"] = round(infer_s * 1000.0, 0)
+
+        pth = threading.Thread(target=planner, daemon=True)
+        pth.start()
+
+        frames = []
+        success = False
+        next_t = time.perf_counter()
+        step = 0
+        while not STOP["flag"] and not shared["done"] and step < max_steps:
+            with buf_lock:
+                action = buffer.popleft() if buffer else None
+                depth_buf = len(buffer)
+            holding = action is None
+            if holding:
+                action = np.zeros((sc.env.num_envs, 7), dtype=np.float32)
+                action[:, 6] = shared["last_gripper"]
+                shared["hold_steps"] += 1
+            else:
+                shared["last_gripper"] = float(action[0, 6])
+            shared["total_steps"] += 1
+
+            try:
+                obs, reward, terminated, truncated, info = sc.env.step(action)
+            except ValueError as e:
+                # robosuite raises "executing action in terminated episode" when the
+                # underlying env set its internal done but the gym wrapper still reported
+                # terminated/truncated False; treat it as a clean episode end instead of
+                # letting the exception kill the engine thread (which froze the demo).
+                print("step after termination, ending episode:", e, flush=True)
+                shared["done"] = True
+                break
+            with LOCK:
+                shared["obs"] = obs
+            if "final_info" in info:
+                try:
+                    success = bool(np.asarray(info["final_info"]["is_success"]).any())
+                except Exception:
+                    pass
+            done = bool(np.any(terminated) or np.any(truncated))
+
             rgb = sc.hi_res(VIEW_RES)
             _set_frame(rgb)
-            frames.append(_banner_frame(rgb, instruction, VIDEO_RES))
-            v = getattr(policy, "_last_model_inference_s", 0.0) * 1000.0
+            frames.append(_banner_frame(rgb, instruction, VIDEO_RES, holding))
+            step += 1
+            hp = 100.0 * shared["hold_steps"] / max(1, shared["total_steps"])
             with LOCK:
-                STATE["step"] += 1
-                if v > 0:  # only replan steps run the model; keep the last real timing
-                    STATE["infer_ms"] = round(v, 0)
+                STATE.update(step=step, holding=holding, buffer=depth_buf, hold_pct=round(hp, 0),
+                             status=("thinking (holding pose)" if holding else f"running: {instruction}"))
+            if done:
+                shared["done"] = True
 
-        success = False
-        try:
-            with torch.no_grad():
-                out = ev.rollout(sc.env, policy, env_preprocessor=env_pre, env_postprocessor=env_post,
-                                 preprocessor=preprocessor, postprocessor=postprocessor,
-                                 seeds=[SEED], render_callback=render_cb)
-            try:
-                success = bool(np.asarray(out["success"]).any())
-            except Exception:
-                success = False
-        except StopRollout:
-            with LOCK:
-                STATE["status"] = "stopped"
+            next_t += DT
+            sleep = next_t - time.perf_counter()
+            if sleep > 0:
+                time.sleep(sleep)
+            else:
+                next_t = time.perf_counter()  # fell behind real-time; resync
 
-        # save debug video (command banner already on each frame)
+        shared["done"] = True
+        STOP["flag"] = True
+        pth.join(timeout=5.0)
+
         url = ""
         if frames:
             ts = datetime.now().strftime("%H%M%S")
-            name = f"interactive/{ts}_{sc.suite}_{sc.task_id}_{'ok' if success else 'run'}.mp4"
+            name = f"interactive_rt/{ts}_{sc.suite}_{sc.task_id}_{'ok' if success else 'run'}.mp4"
             path = os.path.join(OUT_DIR, name)
             try:
                 import imageio
-                with imageio.get_writer(path, fps=20, codec="libx264", quality=8,
+                with imageio.get_writer(path, fps=int(RT_HZ), codec="libx264", quality=8,
                                         macro_block_size=1, output_params=["-pix_fmt", "yuv420p"]) as w:
                     for fr in frames:
                         w.append_data(fr)
@@ -242,9 +344,29 @@ def rollout_thread():
                 print("video save failed:", e, flush=True)
 
         with LOCK:
-            STATE.update(mode="idle", success=success, video_url=url,
+            STATE.update(mode="idle", success=success, video_url=url, holding=False,
                          status=("success" if success else ("stopped" if STOP["flag"] else "done")))
         _set_frame(sc.idle_frame(VIEW_RES))
+
+    # Warm up flash/JIT kernels NOW (the first model forward compiles them and can take
+    # ~20s); otherwise that one-time stall would burn the whole first real-time episode.
+    with LOCK:
+        STATE["status"] = "warming up GPU kernels (one-time JIT) ..."
+    try:
+        wobs, _ = scene.env.reset(seed=SEED)
+        wproc = preprocess_observation(wobs)
+        wproc["task"] = [scene.scene_task or "pick up the object" for _ in range(scene.env.num_envs)]
+        wproc = env_pre(wproc)
+        wproc = preprocessor(wproc)
+        wgen = ev._make_rollout_action_generator(policy, [SEED])
+        _t = time.perf_counter()
+        plan_chunk(wproc, wgen)
+        policy.reset()
+        print(f"[warmup] done in {time.perf_counter()-_t:.1f}s", flush=True)
+    except Exception as e:  # noqa: BLE001
+        import traceback; traceback.print_exc()
+        print("warmup failed:", e, flush=True)
+    show_idle(scene)
 
     # main control loop
     while True:
@@ -266,37 +388,48 @@ def rollout_thread():
                     STATE["status"] = f"scene build failed: {e}"
                 continue
             try:
-                scene.env.close()  # free the old EGL/MuJoCo context
+                scene.env.close()
             except Exception:
                 pass
             scene = new_scene
             show_idle(scene)
         elif action == "run":
-            run_command(scene, instruction)
+            try:
+                run_command(scene, instruction)
+            except Exception as e:  # noqa: BLE001
+                import traceback; traceback.print_exc()
+                with LOCK:
+                    STATE.update(mode="idle", holding=False, status=f"error: {e}")
+                try:
+                    show_idle(scene)
+                except Exception:
+                    pass
 
 
 # ---------------------------------------------------------------------------
 PAGE = b"""<!doctype html><html><head><meta charset=utf-8>
-<title>MolmoAct2 x LIBERO (live)</title>
+<title>MolmoAct2 x LIBERO (REAL-TIME)</title>
 <style>
  body{background:#0f1012;color:#e8e8ea;font-family:system-ui,sans-serif;margin:0;padding:20px}
  .wrap{max-width:760px;margin:0 auto}
- h1{font-size:18px;font-weight:600;margin:0 0 12px}
+ h1{font-size:18px;font-weight:600;margin:0 0 4px}
+ .sub{font-size:12px;color:#8b94a0;margin:0 0 12px}
  img{width:640px;height:auto;border-radius:10px;background:#000;display:block}
  .row{display:flex;gap:8px;margin-top:14px}
  input{flex:1;padding:11px;border-radius:8px;border:1px solid #333;background:#1b1b1f;color:#eee;font-size:15px}
  button{padding:11px 16px;border-radius:8px;border:0;color:#fff;font-size:15px;cursor:pointer}
  .send{background:#3b82f6}.stop{background:#ef4444}.rand{background:#8b5cf6}
  .meta{margin-top:12px;font-size:13px;color:#9aa3ad}
+ .think{color:#ffb450;font-weight:600}
  .panel{margin-top:12px;background:#16171b;border:1px solid #26272c;border-radius:10px;padding:12px;font-size:13px}
  .chip{display:inline-block;background:#23252b;border-radius:14px;padding:3px 10px;margin:3px 4px 0 0;color:#cdd3da}
- a{color:#7aa2ff}
- video{width:640px;border-radius:10px;margin-top:10px;background:#000}
+ a{color:#7aa2ff} video{width:640px;border-radius:10px;margin-top:10px;background:#000}
 </style></head><body><div class=wrap>
-<h1>MolmoAct2 x LIBERO - live sim</h1>
+<h1>MolmoAct2 x LIBERO real-time sim</h1>
+<p class=sub>The simulator runs at wall-clock speed; the robot HOLDS its pose while the model thinks, then moves when the next chunk lands.</p>
 <img src="/stream" alt="sim">
 <div class=row>
- <input id=cmd placeholder="type an instruction, then Send (resets the scene and runs it)">
+ <input id=cmd placeholder="type an instruction, then Send (resets the scene and runs it in real time)">
  <button class=send onclick=send()>Send</button>
  <button class=stop onclick=stop()>Stop</button>
  <button class=rand onclick=rnd()>Randomize</button>
@@ -314,15 +447,16 @@ document.getElementById('cmd').addEventListener('keydown',e=>{if(e.key==='Enter'
 let lastVid='';
 async function poll(){
  try{const s=await(await fetch('/status')).json();
-  document.getElementById('meta').textContent='['+s.mode+'] '+s.status+' | step '+s.step+' | last infer '+s.infer_ms+' ms'+(s.instruction?' | running: "'+s.instruction+'"':'');
-  document.getElementById('scene').textContent='Scene: '+s.suite+' / task '+s.task_id+' - "'+s.scene_task+'"';
+  let t='['+s.mode+'] '+s.status+' | step '+s.step+' | last infer '+s.infer_ms+' ms | buffer '+s.buffer+' | hold '+s.hold_pct+'%';
+  const m=document.getElementById('meta'); m.textContent=t; m.className='meta'+(s.holding?' think':'');
+  document.getElementById('scene').textContent='Scene: '+s.suite+' / task '+s.task_id+': "'+s.scene_task+'"';
   document.getElementById('objs').innerHTML='Objects in scene: '+(s.objects||[]).map(o=>'<span class=chip>'+o+'</span>').join('');
   const box=document.getElementById('cmd');
   if(s.mode==='idle'&&!box.value&&s.scene_task)box.value=s.scene_task;
   if(s.video_url&&s.video_url!==lastVid){lastVid=s.video_url;
    document.getElementById('vidwrap').innerHTML='<div style=\"margin-top:8px;font-size:13px;color:#9aa3ad\">last run video:</div><video controls autoplay loop src=\"'+s.video_url+'\"></video>';}
  }catch(e){}
- setTimeout(poll,800);
+ setTimeout(poll,500);
 }
 poll();
 </script></body></html>"""
@@ -347,7 +481,7 @@ class Handler(BaseHTTPRequestHandler):
             with LOCK:
                 s = {k: STATE[k] for k in ("mode", "status", "instruction", "scene_task",
                                            "suite", "task_id", "objects", "step", "infer_ms",
-                                           "success", "video_url")}
+                                           "success", "holding", "buffer", "hold_pct", "video_url")}
             self._send(200, "application/json", json.dumps(s).encode())
         elif path == "/video":
             with LOCK:
@@ -370,7 +504,7 @@ class Handler(BaseHTTPRequestHandler):
                         self.wfile.write(f"Content-Length: {len(frame)}\r\n\r\n".encode())
                         self.wfile.write(frame)
                         self.wfile.write(b"\r\n")
-                    time.sleep(0.06)
+                    time.sleep(0.04)
             except (BrokenPipeError, ConnectionResetError):
                 pass
         else:
@@ -382,7 +516,7 @@ class Handler(BaseHTTPRequestHandler):
             n = int(self.headers.get("Content-Length", "0"))
             instr = parse_qs(self.rfile.read(n).decode()).get("instruction", [""])[0].strip()
             if instr:
-                STOP["flag"] = True  # abort any running task, then run the new one
+                STOP["flag"] = True
                 with LOCK:
                     PENDING["action"], PENDING["instruction"] = "run", instr
                 EVENT.set()
@@ -405,9 +539,9 @@ class Handler(BaseHTTPRequestHandler):
 
 
 def main():
-    threading.Thread(target=rollout_thread, daemon=True).start()
+    threading.Thread(target=engine_thread, daemon=True).start()
     srv = ThreadingHTTPServer(("0.0.0.0", PORT), Handler)
-    print(f"interactive demo on http://localhost:{PORT}  (remote box: ssh -L {PORT}:localhost:{PORT} <host>)", flush=True)
+    print(f"real-time demo on http://localhost:{PORT}  (remote box: ssh -L {PORT}:localhost:{PORT} <host>)", flush=True)
     srv.serve_forever()
