+++ packages/vla/molmoact2/scripts/model_smoke.py
@@ -1,15 +1,11 @@
 # Copyright(C) 2026 Advanced Micro Devices, Inc. All rights reserved.
 # SPDX-License-Identifier: MIT
-"""Capability 1 — full-model smoke test for MolmoAct2 on AMD Strix Halo (gfx1151).
+"""Full-model smoke test: load MolmoAct2-DROID and run one action prediction on ROCm.
 
-Initializes the *real* MolmoAct2-DROID checkpoint (downloads ~22 GB once into the
-mounted HF cache), then runs a single action-prediction on a dummy observation to
-prove the whole flow-matching action path executes end-to-end on ROCm. Reuses the
-validated upstream `Policy` loader (bf16 dtype patches) baked at
-/repos/molmoact2/examples/droid.
+Downloads ~22 GB once into the HF cache, then predicts on a dummy observation using the
+upstream Policy loader at /repos/molmoact2/examples/droid. Exits non-zero on failure.
 
-Env: MODEL_REPO (allenai/MolmoAct2-DROID), DTYPE (bfloat16), NUM_STEPS (10).
-Exits non-zero on any failure so `ryzers run` / CI catches a broken image.
+Env: MODEL_REPO, DTYPE (bfloat16), NUM_STEPS (10).
 """
 import os
 import sys
@@ -27,8 +23,8 @@ NUM_STEPS = int(os.environ.get("NUM_STEPS", "10"))
 
 
 def predict_chunk(policy, norm_tag, images, task, state, num_steps):
-    """Call the model's flow-matching action head, tolerating the
-    inference_action_mode (new) vs action_mode (old) kwarg rename."""
+    """Call the flow-matching action head, tolerating the inference_action_mode
+    vs action_mode kwarg rename across upstream versions."""
     kw = dict(processor=policy.processor, images=images, task=task, state=state,
               norm_tag=norm_tag, enable_depth_reasoning=False, num_steps=num_steps,
               normalize_language=True, enable_cuda_graph=False)
@@ -62,7 +58,7 @@ def main() -> int:
     from host_server_droid import Policy, NORM_TAG, REPO_ID  # noqa: E402
 
     repo = MODEL_REPO or REPO_ID
-    print(f"loading model    : {repo}  (dtype={DTYPE}) — first run downloads ~22 GB")
+    print(f"loading model    : {repo}  (dtype={DTYPE}); first run downloads ~22 GB")
     t0 = time.time()
     policy = Policy(repo_id=repo, device="cuda:0", dtype=DTYPE)
     n_params = sum(p.numel() for p in policy.model.parameters())
