+++ packages/vla/molmoact2/test.py
@@ -1,11 +1,9 @@
 # Copyright(C) 2026 Advanced Micro Devices, Inc. All rights reserved.
 # SPDX-License-Identifier: MIT
-"""Environment sign-of-life for the MolmoAct2 Ryzer image on Strix Halo (gfx1151).
+"""Environment check for the MolmoAct2 image on Strix Halo (gfx1151).
 
-Runs inside the built image. No model weights required — proves the container
-has (1) a working ROCm torch on the iGPU and (2) the MolmoAct2 server deps at
-the pinned versions, before we pull the ~22 GB checkpoints. Exits non-zero on
-any failure so `ryzers run` / CI catches a broken image early.
+Verifies ROCm torch on the iGPU and the inference deps at pinned versions, with no
+model weights. Exits non-zero on any failure.
 """
 import sys
 
@@ -31,7 +29,6 @@ def main() -> int:
     b = torch.randn(512, 512, device="cuda")
     print(f"matmul ok        : sum={(a @ b).sum().item():.3f}")
 
-    # MolmoAct2 server deps (pins from allenai/molmoact2 pyproject.toml).
     import transformers
     import accelerate
     import huggingface_hub
